Remove `putility.libs.time` and inline time utilities
