+++ src/backend/src/Extension.js
@@ -21,7 +21,28 @@ const { AdvancedBase } = require('@heyputer/putility');
 const EmitterFeature = require('@heyputer/putility/src/features/EmitterFeature');
 const { Context } = require('./util/context');
 const { ExtensionServiceState } = require('./ExtensionService');
-const { display_time } = require('@heyputer/putility/src/libs/time');
+
+const module_epoch_d = new Date();
+const display_time = (now) => {
+    const pad2 = n => String(n).padStart(2, '0');
+
+    const yyyy = now.getFullYear();
+    const mm   = pad2(now.getMonth() + 1);
+    const dd   = pad2(now.getDate());
+    const HH   = pad2(now.getHours());
+    const MM   = pad2(now.getMinutes());
+    const SS   = pad2(now.getSeconds());
+    const time = `${HH}:${MM}:${SS}`;
+
+    const needYear  = yyyy !== module_epoch_d.getFullYear();
+    const needMonth = needYear || (now.getMonth() !== module_epoch_d.getMonth());
+    const needDay   = needMonth || (now.getDate() !== module_epoch_d.getDate());
+
+    if ( needYear ) return `${yyyy}-${mm}-${dd} ${time}`;
+    if ( needMonth ) return `${mm}-${dd} ${time}`;
+    if ( needDay ) return `${dd} ${time}`;
+    return time;
+};
 
 let memoized_errors = null;
