+++ src/putility/index.js
@@ -18,7 +18,6 @@ module.exports = {
         listener: require('./src/libs/listener'),
         log: require('./src/libs/log'),
         string: require('./src/libs/string'),
-        time: require('./src/libs/time'),
         event: require('./src/libs/event'),
     },
     features: {
